deleting files

+++ controllers/admin.js
@@ -1,6 +1,6 @@
 const Product = require('../models/product');
-
-const {validationResult} = require('express-validator')
+const fileHelper = require('../util/file');
+const {validationResult} = require('express-validator');
 
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
@@ -142,6 +142,7 @@ exports.postEditProduct = (req, res, next) => {
     product.description = updatedDescription;
 
     if(image){
+      fileHelper.deleteFile(product.imageUrl);
       product.imageUrl = image.path;
     }
 
@@ -179,7 +180,16 @@ exports.getProducts = (req, res, next) => {
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
 
-  Product.deleteOne({_id: prodId, userId: req.user._id} )
+  Product.findById(prodId)
+  .then(product => {
+
+    if(!product){
+      return next(new Error('Product not found!'))
+    }
+
+    fileHelper.deleteFile(product.imageUrl);
+    return Product.deleteOne({_id: prodId, userId: req.user._id} )
+  })
   .then( () => {
     console.log('Product Deleted');
     res.redirect('/admin/products');
